move segwit buttons into another component
